requirements.txt
beautifulsoup4==4.12.0
google_api_python_client==2.83.0
google_auth_oauthlib==1.0.0
protobuf==4.22.1
requests==2.28.2
